+++ src/TimeTest.php
@@ -192,7 +192,7 @@ final class TimeTest extends TestCase
     public function testAddTime(): void
     {
         $time = Time::at(10)
-            ->add(Duration::of(hours: 2, minutes: 30, seconds: 15, microseconds: 500));
+            ->shift(Duration::of(hours: 2, minutes: 30, seconds: 15, microseconds: 500));
 
         self::assertSame(12, $time->hour);
         self::assertSame(30, $time->minute);
@@ -202,7 +202,7 @@ final class TimeTest extends TestCase
 
     public function testAddTimeWrapsDay(): void
     {
-        $time = Time::at(23)->add(Duration::of(hours: 2));
+        $time = Time::at(23)->shift(Duration::of(hours: 2));
 
         self::assertSame(1, $time->hour);
     }
@@ -211,7 +211,7 @@ final class TimeTest extends TestCase
     {
         $time = Time::at(23);
 
-        self::assertSame($time, $time->add(Duration::of()));
+        self::assertSame($time, $time->shift(Duration::of()));
     }
 
     /* -------------------------------------------------
@@ -262,7 +262,7 @@ final class TimeTest extends TestCase
     {
         $time = Time::at(10);
 
-        self::assertSame($time, $time->add(Duration::zero()));
+        self::assertSame($time, $time->shift(Duration::zero()));
     }
 
     public function testMicroseconds(): void
@@ -444,7 +444,7 @@ final class TimeTest extends TestCase
 
         $this->expectExceptionObject(InvalidDuration::dueToOverflow());
 
-        $time->add(Duration::max());
+        $time->shift(Duration::max());
     }
 
     public function test_time_can_be_serialized_and_unserialized(): void